refactor: Extract pure types and add FileSystem abstraction
Code reorganization to improve modularity:

- Extract Action and KeyContext types from keybindings.rs to action.rs
- Move CursorId from event.rs to cursor.rs
- Add FileSystem trait for abstracting file I/O operations
- Refactor Buffer to use FileSystem trait for portability
- Add GrammarStore injection to GrammarRegistry for flexibility
- Refactor HighlightEngine for better separation of concerns
- Extract shared Language enum from highlighter.rs
- Extend FsBackend trait with additional operations
- Improve theme portability
