+++ src/model/event.rs
@@ -1,18 +1,10 @@
+// Re-export CursorId from cursor module for backwards compatibility
+pub use crate::model::cursor::CursorId;
 use crate::model::piece_tree::PieceTree;
 use crate::view::overlay::{OverlayHandle, OverlayNamespace};
 use serde::{Deserialize, Serialize};
 use std::ops::Range;
 use std::sync::Arc;
-
-/// Unique identifier for a cursor
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct CursorId(pub usize);
-
-impl CursorId {
-    /// Sentinel value used for inverse events during undo/redo
-    /// This indicates that the event shouldn't move any cursor
-    pub const UNDO_SENTINEL: CursorId = CursorId(usize::MAX);
-}
 
 /// Unique identifier for a split pane (re-exported from split.rs)
 /// Note: This is defined in split.rs and re-exported here for events
@@ -570,7 +562,8 @@ pub struct EventLog {
     /// How often to create snapshots (every N events)
     snapshot_interval: usize,
 
-    /// Optional file for streaming events to disk
+    /// Optional file for streaming events to disk (runtime-only, for debugging)
+    #[cfg(feature = "runtime")]
     stream_file: Option<std::fs::File>,
 
     /// Index at which the buffer was last saved (for tracking modified status)
@@ -586,6 +579,7 @@ impl EventLog {
             current_index: 0,
             snapshots: Vec::new(),
             snapshot_interval: 100,
+            #[cfg(feature = "runtime")]
             stream_file: None,
             saved_at_index: Some(0), // New buffer starts at "saved" state (index 0)
         }
@@ -621,7 +615,8 @@ impl EventLog {
         }
     }
 
-    /// Enable streaming events to a file
+    /// Enable streaming events to a file (runtime-only, for debugging)
+    #[cfg(feature = "runtime")]
     pub fn enable_streaming<P: AsRef<std::path::Path>>(&mut self, path: P) -> std::io::Result<()> {
         use std::io::Write;
 
@@ -641,12 +636,14 @@ impl EventLog {
         Ok(())
     }
 
-    /// Disable streaming
+    /// Disable streaming (runtime-only)
+    #[cfg(feature = "runtime")]
     pub fn disable_streaming(&mut self) {
         self.stream_file = None;
     }
 
-    /// Log rendering state (for debugging)
+    /// Log rendering state (for debugging, runtime-only)
+    #[cfg(feature = "runtime")]
     pub fn log_render_state(
         &mut self,
         cursor_pos: usize,
@@ -674,7 +671,8 @@ impl EventLog {
         }
     }
 
-    /// Log keystroke (for debugging)
+    /// Log keystroke (for debugging, runtime-only)
+    #[cfg(feature = "runtime")]
     pub fn log_keystroke(&mut self, key_code: &str, modifiers: &str) {
         if let Some(ref mut file) = self.stream_file {
             use std::io::Write;
@@ -702,14 +700,15 @@ impl EventLog {
             self.entries.truncate(self.current_index);
         }
 
-        // Stream event to file if enabled
+        // Stream event to file if enabled (runtime-only)
+        #[cfg(feature = "runtime")]
         if let Some(ref mut file) = self.stream_file {
             use std::io::Write;
 
             let stream_entry = serde_json::json!({
                 "index": self.entries.len(),
                 "timestamp": chrono::Local::now().to_rfc3339(),
-                "event": event,
+                "event": &event,
             });
 
             // Write JSON line and flush immediately for real-time logging
@@ -837,7 +836,8 @@ impl EventLog {
         self.snapshots.clear();
     }
 
-    /// Save event log to JSON Lines format
+    /// Save event log to JSON Lines format (runtime-only)
+    #[cfg(feature = "runtime")]
     pub fn save_to_file(&self, path: &std::path::Path) -> std::io::Result<()> {
         use std::io::Write;
         let file = std::fs::File::create(path)?;
@@ -851,7 +851,8 @@ impl EventLog {
         Ok(())
     }
 
-    /// Load event log from JSON Lines format
+    /// Load event log from JSON Lines format (runtime-only)
+    #[cfg(feature = "runtime")]
     pub fn load_from_file(path: &std::path::Path) -> std::io::Result<Self> {
         use std::io::BufRead;
         let file = std::fs::File::open(path)?;
